# auth-service/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  auth-service_internal_domain.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      picture:
        description: Allows NULL
        type: string
      role:
        description: '"user"'
        type: string
      status:
        description: '"unverified", "verified"'
        type: string
      updated_at:
        type: string
    type: object
  auth-service_internal_web_helpers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  internal_user.UpdateUserParams:
    properties:
      name:
        type: string
      picture:
        type: string
      role:
        type: string
      status:
        type: string
    type: object
host: localhost:5000
info:
  contact: {}
  description: This is the API for the HealthMate application.
  title: HealthMate API
  version: "1.0"
paths:
...
  /users:
    get:
      description: Retrieve a list of users with optional search and pagination
      parameters:
      - description: Search query
        in: query
        name: search
        type: string
      - description: Limit (default 20)
        in: query
        name: limit
        type: integer
      - description: Offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/auth-service_internal_domain.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/auth-service_internal_web_helpers.ErrorResponse'
      security:
      - Bearer: []
      summary: List/Search users
      tags:
      - users
  /users/profile:
    get:
      description: Get profile details of the currently authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth-service_internal_domain.User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth-service_internal_web_helpers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/auth-service_internal_web_helpers.ErrorResponse'
      security:
      - Bearer: []
      summary: Get user profile
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update profile details (name, picture) of the current user
      parameters:
      - description: Profile updates
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_user.UpdateUserParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth-service_internal_web_helpers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth-service_internal_web_helpers.ErrorResponse'
      security:
      - Bearer: []
      summary: Update user profile
      tags:
      - users
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"